feat(frontend): re-use BTC address listener service
