+++ src/frontend/src/icp/services/worker.ckbtc-update-balance.services.ts
@@ -1,11 +1,8 @@
-import {
-	syncBtcAddress,
-	syncBtcPendingUtxos,
-	syncCkBTCUpdateOk
-} from '$icp/services/ckbtc-listener.services';
+import { syncBtcPendingUtxos, syncCkBTCUpdateOk } from '$icp/services/ckbtc-listener.services';
 import { btcAddressStore } from '$icp/stores/btc.store';
 import type { IcCkWorker, IcCkWorkerInitResult, IcCkWorkerParams } from '$icp/types/ck-listener';
 import { isNetworkIdBTCMainnet } from '$icp/utils/ic-send.utils';
+import { syncBtcAddress } from '$lib/services/listener.services';
 import type {
 	PostMessage,
 	PostMessageDataResponseBTCAddress,
